02_Project_Vote/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
Werkzeug==3.0.3
python-dotenv==1.0.1
email-validator==2.1.0.post1
gunicorn==21.2.0
pytest==8.2.1
